Update to Mono V2
https://discord.com/channels/882234192132448256/1120390177613942836/1171608349075312680

+++ ScaleformUI_Csharp/Scaleforms/BigMessage/BigMessage.cs
@@ -1,4 +1,5 @@
 ﻿using CitizenFX.Core;
+using CitizenFX.FiveM;
 using ScaleformUI.Elements;
 
 namespace ScaleformUI.Scaleforms
@@ -41,7 +42,7 @@ namespace ScaleformUI.Scaleforms
             if (ManualDispose)
             {
                 _sc.CallFunction(Transition, TransitionDuration, TransitionAutoExpansion);
-                await BaseScript.Delay((int)((TransitionDuration * .5f) * 1000));
+                await BaseScript.Delay((uint)(int)((TransitionDuration * .5f) * 1000));
                 ManualDispose = false;
             }
